+++ crypto_history/__init__.py
@@ -1,4 +1,5 @@
 from .utilities.general_utilities import init_logger, AbstractFactory  # NOQA
-from .core import stock_market_factory, data_container_pre  # NOQA
+from .data_container import data_container_access  # NOQA
+from .stock_market import stock_market_factory # NOQA
 
 class_builders = AbstractFactory.get_builders()