+++ crypto_history/utilities/crypto_enum.py
@@ -1,1 +1,0 @@
-UNKNOWN_REFERENCE_COIN = "unknown_reference_coin"